requirements.txt
# VR Video Processing Pipeline - Core Dependencies
# 这是基于实际代码分析生成的精简依赖列表

# Core configuration and system monitoring
PyYAML>=6.0.0
psutil>=5.9.0
GPUtil>=1.4.0

# Video processing (FFmpeg wrapper)
ffmpeg-python>=0.2.0

# Optional: Enhanced CLI experience
click>=8.0.0
rich>=12.0.0

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Note: This project primarily relies on external tools like FFmpeg
# Most video processing is done through subprocess calls to FFmpeg
# Additional dependencies can be added as needed for specific features
